(main): Reformat if-block a little. Declare index I to be of type size_t to avoid warning from gcc -Wall.

+++ src/stty.c
@@ -920,14 +920,14 @@ main (int argc, char **argv)
 	  if (speed_was_set || memcmp (&mode, &new_mode, sizeof (mode)) != 0)
 #endif
 	    {
-	    int i;
-	    error (1, 0,
+	      size_t i;
+	      error (1, 0,
 	      _("standard input: unable to perform all requested operations"));
-	    printf (_("new_mode: mode\n"));
-	    for (i=0; i<sizeof(new_mode); i++)
-	      printf ("0x%02x: 0x%02x\n",
-		      *(((unsigned char *) &new_mode) + i),
-		      *(((unsigned char *) &mode) + i));
+	      printf (_("new_mode: mode\n"));
+	      for (i = 0; i < sizeof (new_mode); i++)
+		printf ("0x%02x: 0x%02x\n",
+			*(((unsigned char *) &new_mode) + i),
+			*(((unsigned char *) &mode) + i));
 	    }
 	}
     }
